lantiq: BT Home Hub 2B support - nand driver locking
As a result of changeset 40310, the xway nand driver
now acquires ebu_lock in the chip select function, and
holds it for the entire duration of an operation until
the chip is deselected. There is surely no longer any
need therefore also to acquire the lock in each separate
read or write function. This patch removes that code.

Signed-off-by: Ben Mulvihill <ben.mulvihill@gmail.com>